Fix for issue 99594 (Scientific notations not supported as XPath literals)

+++ xml.xpath/src/org/netbeans/modules/xml/xpath/impl/XPathTreeCompiler.java
@@ -45,7 +45,7 @@ public class XPathTreeCompiler extends TreeCompiler {
 	public Object number(String value) {
 		//distinguish between Long and Double 
 		try {
-			int intVal = Integer.parseInt(value);
+			long intVal = Long.parseLong(value);
 			return new Constant(new Long(value));
 		} catch(NumberFormatException ex) {
 			//Do Nothing
